fanotify: add a super block root watch
Set a watch on the super block's root inode including the
mask bit FAN_EVENT_ON_SB to get notified on the events
of all the inodes on the same super block.

When requesting to add a super block root watch, any
file on the file system can be passed as input argument
for fanotify_mark(). The mark will be added to the root
inode of the file system that file belongs to.

The super block watch cannot be set with FAN_MARK_MOUNT flag,
because a super block watch is an inode mark, not a mount mark.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>