ovl: redirect dir on rename in snapshot mount
Directory rename in snapshot mount sets the redirect xattr
on the upper path directory. This is needed is order to
keep a reference to the snapshot dir where children files
should be copied to.

Redirect must be by path and not by file handle, because
the underlying overlayfs mount does not support exportfs ops.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>