ovl: follow mismatch stored origin fh of merge dir
When the 'verify' feature is enabled, if lower dir does not match the
origin fh stored in upper dir, or if no lower dir was found by name,
try to decode the stored origin fh and use the result as the merge dir
lower dentry.

When there is more than a single lower layer, a lower merge directory
followed from upper dir by origin fh is not followed down again to lower
layers by name. This will be fixed by upcoming patch.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>