ovl: introduce snapshot stacked filesystem
Register overlayfs module also as filesystem type "snapshot" that
uses the same inode/file operation as an "overlay" filesystem.

Like an overlay filesystem, a snapshot filesystem is stacked on top
of an underlying filesystem and all the operations on the snapshot
filesystem objects pass through to the underlying filesystem objects.

Later on, the snapshot filesystem will provide copy on write to
a snapshot storage directory.

The snapshot filesystem reuses a lot of overlayfs code, but has no
lowerdir and there is no copy up, so there is no need for a workdir.
Fix some places in overlayfs code that assume the existence of lowerdir
and workdir.

The snapshot filesystem is enabled by the experimental config
option CONFIG_OVERLAY_FS_SNAPSHOT.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>

fixintro