ovl: implement snapshot fs lookup
Snapshot lookup is about finding out if a file has already been
copied up or whited out in the snapshot overlay.

On lookup of a snapshot fs path, we find a dentry in the snapshot
overlay fs that refers back to the snapshot fs underlying path or
is a copy up of the upderlying path.

When looking up in the snapshot overlay fs, all we need to know
about the snaphot overlay dentry is if it is positive and if it is
upper, so we know if it was already copied to snapshot.

If we find that the object doesn't need to be copied to snapshot or
that its children (i.e. if it is a directory) don't need to be copied,
we mark the result in snapshot fs dentry flags.

While looking up in snapshot overlay, we need to hold a reference on the
snapshto overlay mount, because the snapshot overlay may be unmounted on
snapshot take, but we cannot hold a reference to the found snapshot
overlay dentry in the snapshot fs dentry, because the snapshot overlay
dentries will be evicted on snapshot take.

Instead, we rely on snapshot fs lookup to populate the dentry and inode
cache of the snapshot overlay, so we can quickly find the snapshot
overlay dentry by the snapshot fs underlying dentry when we need it.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>