ovl: add support for verify_dir mount option
When overlayfs is mounted with option 'verify_dir', a directory inode
found in lower layer by name or by redirect_dir is verified against the
file handle of the copy up origin that is stored in the upper layer.
If origin xattr does not exist, update it with the lower found by name.

This option introduces a change of behavior for the case of lower layer
modification while overlay is offline. A lower directory created or
moved offline under an exisitng upper directory, will not be merged with
that upper directory when the overlay in mounted with 'verify_dir'.

The 'verify_dir' option should not be used after copying layers,
because the new lower directory inodes would fail verification.

This is going to be used for overlayfs snapshots and NFS export.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>