ovl: fix read-only mount on unsupported index entries
The index=on feature in kernel v4.13 did not support directory
and whiteout index file types.

When encoutering an index entry of type directory or whiteout,
kernel it was supposed to fallback to read-only mount, but the
fill_super() operation returns EROFS instead of returning success
with read-only mount flag, so mount fails when encoutering
directory or whiteout index entries.

Fixes: 61b674710cd9 ("ovl: do not cleanup directory and whiteout index..")
Cc: <stable@vger.kernel.org> # v4.13
Signed-off-by: Amir Goldstein <amir73il@gmail.com>