ovl: whiteout orphan index entries on mount
Invalid and stale index entries are removed on mount.

Orphan index entries are non-dir index entries whose union nlink count
dropped to zero. With index=on, orphan index entries are removed on
mount. With index=all orphan index entries are replaced with white out
index entries to block future open by handle from opening the lower
file.

When dir index has a stale origin fh to upper dir, we assume that the
upper dir was removed and we treat the dir index as orphan entry that
needs to be whited out or removed.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>