ovl: set origin xattr for merge dir on lookup
For old existing merge dirs whose "origin" xattr was not set at copy up
time, we amend the situation on lookup.

If no origin fh is stored in upper of a merge dir, store fh of upper most
lower dir or 'null' fh if lower does not support file handles. We do this
so we can filter out whiteouts in case lower dir is removed offline and
then upper dir becomes a pure upper in a future mount.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>