ovl: copy up any file on access with "migrate" feature
If the "migrate" mount option is set then overlay filesystems will copy up
any file and directory on first access.

With this feature, it is possible to migrate all files from source
directory tree to to target directory tree lazily, by using the source
directory as the overlayfs lowerdir and target directory as the
overlayfs upperdir, and then running a 'ls -lR' command on the overlayfs
mount.

The overlayfs mount can be accessed and modified while the migration
is in progress.

The copy on access behavior introduces a slight change in d_real()
semantics. Now d_real() may return an error with NULL inode argument
also in the zero open flags case. vfs API documentation has been updated.

This change does not copy up files on lookup, so non-upper overlay dentries
will still exist in dcache and may still be accesses without copy up for
inode operations, such as permission check and atime update. Copy up on
lookup with be handled by a following patch.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>