ovl: implement metadata only copy up
With consistent fd feature, if clone-up is not supported, we copy up
only the inode metadata. Lower and upper inode pages are expected
to be shared and copied on write lazily.

For now, enabling consistent fd feature without clone support will
cause data corruption of regular files on copy up.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>