ovl: cow file to snapshot on open for write
On open for write of file in a snapshot mount, call ovl_copy_up() on the
snapshot overlay dentry, which will "copy down" the file to the snapshot
overlay's upper dir.

Then, ovl_open() will open a file with its own upper dentry (not the
snapshot overlay's upper dentry), so the stable copy of the file will
remain in the snapshot overlay upper dir.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>