ovl: set opaque xattr if origin fh is stale
With mount option "redirect_dir=origin", if upper dir has a stored origin
fh xattr and that file handle is stale or points to a directory that was
moved outside of the lower layer root, that upper dir becomes implicitly
opaque.

Decoding a directory file handle can be expensive, because it involves
reading directories up to root of the super block and trying to connect
the decoded dentry.

When a stale origin is detected, set the opaque xattr on upper dir to
avoid trying to decode the stale origin fh on the next lookup.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>