ovl: unbless lower st_ino of files under unverified redirected dir
On a malformed overlay, several redirected dirs can point to the same
dir on a lower layer. This presents a similar challenge as broken
hardlinks, because different objects in the overlay can return the same
st_ino/st_dev pair from stat(2).

For broken hardlinks, we do not provide constant st_ino on copy up to
avoid this inconsistency. When "verify" feature is enabled, apply
the same logic to files nested under unverified redirected dirs.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>