ovl: set redirect xattr for lower followed by origin fh
With mount option "redirect_dir=origin", if upper dir has a stored
origin fh xattr and that file handle points to a lower directory that
was renamed, the upper dir becomes implicitly redirected.

Decoding a directory file handle can be expensive, because it involves
reading directories up to root of the super block and trying to connect
the decoded dentry.

When a renamed origin is detected, set the redirect xattr on upper dir
to the path of the decoded origin, so future lookups will find the
correct lower.

If there are lower layers below the decoded origin, return -ESTALE after
setting upper redirect xattr to repeat lookup again by redirect and
continue to lower layers lookup.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>