ovl: add support for 'verify' feature
Introduce the 'verify' config, module and mount options.

If the 'verify' feature is enabled then overlay filesystems will use
the inodes index dir to verify layers consistency.

It is possible to enable consistency verification by default during
build time with the OVERLAY_FS_VERIFY config option, during runtime
with the "verify=on" module option or on a filesystem instance basis
with the "verify=on" mount option.

The 'verify' feature will prevent multiple redirects to the same lower
dir and will prevent broken hardlinks from using the same inode number.

On an overlay filesystem instance with multiple lower layers, 'verify'
feature can only use the index to prevent multiple redirects from upper
dirs and not from middle layer dirs. Emit a warning about this on mount
with multiple lower layers and 'verify=on'.

This is going to be used for NFS export.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>