ovl: support nested overlay file handles
Prepending another ovl_fh header for nested overlay file handle adds
no new information for decoding, so just set a 'nested' flag in ovl_fh
header for nested file handle, so we know to distinguish between real
upper file handle that should be decoded from real upper layer and
nested upper file handle that should be decode from nested lower layer.
Nesting depth cannot be larger than 1, so one bit is enough.

To make the code a bit simpler, the 'nested' flag is only added 'on the
wire' and not stored in origin xattr of a nested overlay's upper.
The 'nested' flag information is not needed in origin xattr, because
copy up origin inode is never an upper inode and upper origin xattr
of directory index entries is never nested.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>