ovl: index directories on copy up with 'verify=on'
With the 'verify' feature enabled, all dirs are indexed on copy up.
Non-dir files are copied up directly to indexdir and then hardlinked
to upper dir.

Directories are copied up to indexdir, then an index entry is created
in indexdir with origin xattr pointing to the copied up dir and then
the copied up dir is moved to upper dir.

This is going to be used for NFS export.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>