ovl: copy up before encoding connectable non-dir file handle
Decoding a connectable overlay non-dir from origin file handle is not
implemented, so we should not encode a connectable lower file handle,
because if file gets copied up after we encoded it, we won't be able to
decode the file handle.

To mitigate this case, we copy up the lower file first and then encode
an upper file handle.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>