ovl: pass ovl_layer array to ovl_check_origin_fh()
Pass the fs instance lower_layer array instead of lowerstack array to
ovl_check_origin_fh(), because the dentry member of ovl_path plays no
part in this helper.

This change simplifies the argument list of ovl_cleanup_index()
and ovl_verify_index().

We pass lower_layer array and num and not struct ovl_fs, because we
are going to use this helper for decoding upper file handles for
NFS export support.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>