ovl: decode connected upper file handles
Until this change, we decoded upper file handles by instantiating an
overlay dentry from the real upper dentry. This is sufficient to handle
pure upper files, but insufficient to handle merge/impure dirs and upper
non-dir with copy up origin.

To that end, if decoded real upper dentry is connected and hashed, we
lookup an overlay dentry with the same path as the real upper dentry.
If decoded real upper is disconnected or unlinked non-dir, we instantiate
a disconnected overlay dentry as before this change.

Because ovl_fh_to_dentry() and ovl_fh_to_parent() methods return a
connected overlay dentry, exportfs does not need to call get_parent()
and get_name() to reconnect an upper overlay dir and only needs to call
get_name() to reconnect a connectable non-dir to its parent. Therefore,
the method ovl_get_parent() should never be called and it is implemented
just to print out a sanity warning.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>