ovl: lookup in inode cache when decoding lower file handle
When decoding a lower file handle, we need to check if lower file was
copied up and indexed and if it has a whiteout index, we need to check
if this is an unlinked but open non-dir before returning -ESTALE.

To find out if this is an unlinked but open non-dir we need to lookup
an overlay inode in inode cache with an alias.

Before this change, the overlay inode was looked up by lower inode and
that required decoding the lower file handle before looking in inode
cache. If the lower inode turned out to be a directory, we may have
paid an expensive cost to reconnect that lower directory for nothing.

After this change, we start by looking up an overlay inode in inode
cache by lower file handle. If we don't find an overlay inode and dentry
alias, then we only need to decode the lower file handle in case the
lower file is not indexed or if the index is a non-dir.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>