fs/namei.c: factor out helper full_name_long_hash()
This is a flavor of full_name_hash() which does not truncate the long
hash value to 32-bit in case long is 64-bit.

A helper __hash_64() is introduced to multiply with GOLDEN_RATIO_64,
just as the the helper __hash_32() multiples with GOLDEN_RATIO_32.
This makes the 32 and 64 bit variants of fold_long_hash() look similar.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>