ovl: opaque xattr should overrule redirect xattr
opaque xattr and redirect xattr are not expected to be both set on the
same directory, but in case they do, the opaque directive to stop
merging lower dirs should overrule an absolute redirect directive to
escape a pure directory.

This change comes with a cost of removing the optimiaztion for not
checking opaque xattr on lookup in last parent layer.
Instead, it adds an optimization for not checking opaque nor redirect
xattr on lookup in overlay instance lowermost layer.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>