ovl: consistent i_ino for non-samefs with xino
When overlay layers are not all on the same fs, but all inode numbers
of underlying fs do not use the high 'xino' bits, overlay st_ino values
are constant and persistent.

In that case, if i_ino member is 64bit, set i_ino value to the same value
as st_ino for nfsd readdirplus validator.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>