xfs: publish max used inode number bits in struct super_block
Publish the max number of inode number bits that could be used after
growing fs to the maximum number of AG. The current inode allocation
mode is not relevant, because we publish the maximal inode number we
could ever encounter in the filesystem instance.

This is going to be used by overlayfs to determine if it create a
unified inode number address space for inodes in all layers of the
overlay when xfs in the underlying filesystem.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>