ovl: constant inode numbers for some nested overlay cases
Constant and persistent inode numbers require that underlying lower
fs supports file handles and that those could be stored in xattr in
upper inodes.

An additional requirement is that lower fs has a non zero sb->s_uuid
field, which identifies the layer in the stored xattr.

A lower overlayfs in a nested overlay mount with nfs_export enabled
meets the file handle support requirement, but doesn't meet the non
zero uuid reuqirement.

Relax the requirement of non zero uuid for lower overlayfs with
nfs_export enabled.  This allows a nested overlay mount to have constant
and persistent inode numbers for non-dir.  In combination with xino for
nested overlay, whose lower overlay is samefs, this can provide
constant and persistent inode numbers for all inodes.

The file handles encoded from the lower overlay are stored in xattr
exactly as lower overlayfs encodes them, without an additional layer
of encapsulation.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>